Fix up errors in tests and test helper functions
The tests for the use of binaries in connectives checked the
truthyness of the outcome, however, in elisp 0 is true so tests that
output zero were failing because we expected them to be false.

The helper functions that get tables were returning new lines at the
end of the string because that is what was in the files. This is not
how the table builder functions operate and not how we want them to,
so use s-chomp to prevent that.